put season score
