+++ src/lib/dynamicData/seasonScore.ts
@@ -1,0 +1,12 @@
+import axios from 'axios';
+
+import type { SeasonScoreDataType } from '@/types/index';
+import { backendUrls } from '@/lib/backendUrls';
+
+export function getSeasonScore(): Promise<number | null> {
+	return axios
+		.get(backendUrls['season_score'])
+		.then((response) => response.data)
+		.then((data: SeasonScoreDataType) => data.limit)
+		.catch(() => null);
+}
